working on cleaning up tests
